(filament_swirl) interlace setup and finalize
