+++ applications/clawpack/advection/2d/filament_swirl/swirl/swirl.cpp
@@ -47,43 +47,3 @@ fclaw2d_domain_t* swirl_create_domain(sc_MPI_Comm mpicomm, fclaw_options_t* gpar
     return domain;
 }
 
-void swirl_run_program(fclaw2d_global_t* glob)
-{
-    char* old_path = fclaw_cwd();
-    fclaw_cd("swirl");
-
-    const swirl_options_t           *swirl_opt;
-
-    /* ---------------------------------------------------------------
-       Set domain data.
-       --------------------------------------------------------------- */
-    fclaw2d_domain_data_new(glob->domain);
-
-    swirl_opt = swirl_get_options(glob);
-
-    /* Initialize virtual table for ForestClaw */
-    fclaw2d_vtables_initialize(glob);
-
-    /* Initialize virtual tables for solvers */
-    if (swirl_opt->claw_version == 4)
-    {
-        fc2d_clawpack46_solver_initialize(glob);
-    }
-    else if (swirl_opt->claw_version == 5)
-    {
-        fc2d_clawpack5_solver_initialize(glob);
-    }
-
-    swirl_link_solvers(glob);
-
-    /* ---------------------------------------------------------------
-       Run
-       --------------------------------------------------------------- */
-    fclaw2d_initialize(glob);
-    fclaw2d_run(glob);
-
-    fclaw2d_finalize(glob);
-
-    fclaw_cd(old_path);
-    FCLAW_FREE(old_path);
-}
